change from float to double

+++ division.h
@@ -3,6 +3,6 @@
 
 #include <stdio.h>
 
-float division(float number1, float number2);
+double division(double number1, double number2);
 
 #endif
